requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.0.0
line-bot-sdk>=3.0.0
anthropic>=0.40.0
supabase>=2.0.0
playwright>=1.48.0
